refactor(database): migrate venue database configuration to new structure
- Updated the environment variable from VENUES_DATABASE_URL to V0_DATABASE_URL for better clarity.
- Refactored database access in get-venue-names.ts to use the new v0-db module.
- Introduced new v0-db and v0-schema files to support the updated database structure.
- Removed the old venues-db file to streamline the database management.
