+++ .env.example
@@ -10,7 +10,7 @@
 NEON_DATABASE_URL="postgresql://username:password@host/database?sslmode=require"
 
 # Venues Database Configuration (separate database for venue data, probably also on neondb)
-VENUES_DATABASE_URL="postgresql://username:password@host/venues_database?sslmode=require"
+V0_DATABASE_URL="postgresql://username:password@host/venues_database?sslmode=require"
 
 # =============================================================================
 # DOMAIN CONFIGURATION
